+++ syscall.c
@@ -40,21 +40,6 @@
 #include "ifuncs.h"
 #include "vfs/vfs_internal.h"
 
-extern int dry_run;
-extern int am_root;
-extern int am_sender;
-extern int read_only;
-extern int list_only;
-extern int inplace;
-extern int preallocate_files;
-extern int sparse_files;
-extern int preserve_perms;
-extern int preserve_executability;
-extern int open_noatime;
-extern int copy_links;
-extern int copy_unsafe_links;
-extern int insecure_links;
-extern int module_id;
 
 
 
@@ -1258,171 +1243,6 @@ int do_chmod_at(const char *fname, mode_t mode)
 }
 #endif
 
-int do_rename(const char *old_path, const char *new_path)
-{
-	if (dry_run) return 0;
-	RETURN_ERROR_IF_RO_OR_LO;
-	return rename(old_path, new_path);
-}
-
-/*
-  Symlink-race-safe variant of do_rename() for receiver-side use. See
-  the comment on do_chmod_at() for the threat model and design rationale.
-
-  rename() is the central tmp -> final operation in rsync; if either the
-  source or the destination has an attacker-substituted symlink in one
-  of its parent components, the rename can publish or vanish files
-  outside the module. Defence: open the parent of *each* path under
-  vfs_resolve_open() and use renameat() against the resulting
-  dirfds. When old_path and new_path share the same parent (the common
-  case -- tmp file living next to its final name), we reuse the same
-  dirfd for both sides.
-
-  Falls through to do_rename() in dry-run, non-daemon, chrooted and
-  absolute-path cases, identical to the other do_*_at() wrappers.
-*/
-int do_rename_at(const char *old_path, const char *new_path)
-{
-#ifdef AT_FDCWD
-	char old_dirpath[MAXPATHLEN], new_dirpath[MAXPATHLEN];
-	const char *old_bname, *new_bname;
-	const char *old_slash, *new_slash;
-	int old_dfd = AT_FDCWD, new_dfd = AT_FDCWD;
-	BOOL old_owns = False, new_owns = False;
-	int ret = -1, e;
-	size_t old_dlen = 0, new_dlen = 0;
-
-	if (dry_run) return 0;
-	RETURN_ERROR_IF_RO_OR_LO;
-
-	if (!vfs_relpath_active())
-		return do_rename(old_path, new_path);
-
-	if (!old_path || !*old_path || !new_path || !*new_path)
-		return do_rename(old_path, new_path);
-
-#if defined O_NOFOLLOW && defined O_DIRECTORY
-	/* Operator-supplied path (e.g. a --backup-dir destination or a --temp-dir
-	 * source): resolve each side's parent via the ownership walk (follow
-	 * uid0/euid symlinks, refuse others; absolute and relative alike). */
-	if (vfs.operator_path_resolve) {
-		if (vfs_symlink_optout_allowed())
-			return do_rename(old_path, new_path);
-		old_dfd = vfs_owner_walk_parent(old_path, &old_bname);
-		if (old_dfd < 0)
-			return -1;
-		new_dfd = vfs_owner_walk_parent(new_path, &new_bname);
-		if (new_dfd < 0) {
-			e = errno;
-			close(old_dfd);
-			errno = e;
-			return -1;
-		}
-		ret = renameat(old_dfd, old_bname, new_dfd, new_bname);
-		e = errno;
-		close(new_dfd);
-		close(old_dfd);
-		errno = e;
-		return ret;
-	}
-#endif
-
-	old_slash = strrchr(old_path, '/');
-	new_slash = strrchr(new_path, '/');
-
-	/* Confine each side independently.  A *relative* side is a transfer path,
-	 * confined beneath the tree via secure_relative_open().  An *absolute* side is
-	 * an operator path (an absolute --temp-dir/--partial-dir temp file): resolve
-	 * its parent via the ownership walk so a flipped foreign-owned parent symlink
-	 * can't redirect the rename out of tree, while still allowing the operator's
-	 * own dirs/".."/uid0-or-euid symlinks.  (--insecure-links keeps the legacy
-	 * unconfined AT_FDCWD path.)  Doing each side independently means an absolute
-	 * source never disables confinement of a relative destination. */
-	if (*old_path == '/') {
-#if defined O_NOFOLLOW && defined O_DIRECTORY
-		if (!symlink_optout_allowed()) {
-			operator_path_resolve = 1;	/* operator side: enforce module-exclude */
-			old_dfd = owner_walk_parent(old_path, &old_bname);
-			operator_path_resolve = 0;
-			if (old_dfd < 0)
-				return -1;
-			old_owns = True;
-		} else
-#endif
-			old_bname = old_path;
-	} else if (old_slash) {
-		old_dlen = old_slash - old_path;
-		if (old_dlen >= sizeof old_dirpath) {
-			errno = ENAMETOOLONG;
-			return -1;
-		}
-		memcpy(old_dirpath, old_path, old_dlen);
-		old_dirpath[old_dlen] = '\0';
-		old_bname = old_slash + 1;
-		old_dfd = vfs_resolve_open(NULL, old_dirpath, O_RDONLY | O_DIRECTORY, 0);
-		if (old_dfd < 0)
-			return -1;
-		old_owns = True;
-	} else {
-		old_bname = old_path;
-	}
-
-	if (*new_path == '/') {
-#if defined O_NOFOLLOW && defined O_DIRECTORY
-		if (!symlink_optout_allowed()) {
-			operator_path_resolve = 1;	/* operator side: enforce module-exclude */
-			new_dfd = owner_walk_parent(new_path, &new_bname);
-			operator_path_resolve = 0;
-			if (new_dfd < 0) {
-				e = errno;
-				if (old_owns) close(old_dfd);
-				errno = e;
-				return -1;
-			}
-			new_owns = True;
-		} else
-#endif
-			new_bname = new_path;
-	} else if (new_slash) {
-		new_dlen = new_slash - new_path;
-		if (new_dlen >= sizeof new_dirpath) {
-			e = ENAMETOOLONG;
-			if (old_owns) close(old_dfd);
-			errno = e;
-			return -1;
-		}
-		memcpy(new_dirpath, new_path, new_dlen);
-		new_dirpath[new_dlen] = '\0';
-		new_bname = new_slash + 1;
-		if (old_owns && old_dlen == new_dlen
-		 && memcmp(old_dirpath, new_dirpath, old_dlen) == 0) {
-			new_dfd = old_dfd;
-		} else {
-			new_dfd = vfs_resolve_open(NULL, new_dirpath, O_RDONLY | O_DIRECTORY, 0);
-			if (new_dfd < 0) {
-				e = errno;
-				if (old_owns) close(old_dfd);
-				errno = e;
-				return -1;
-			}
-			new_owns = True;
-		}
-	} else {
-		new_bname = new_path;
-	}
-
-	ret = renameat(old_dfd, old_bname, new_dfd, new_bname);
-	e = errno;
-	if (new_owns)
-		close(new_dfd);
-	if (old_owns)
-		close(old_dfd);
-	errno = e;
-	return ret;
-#else
-	return do_rename(old_path, new_path);
-#endif
-}
 
 #ifdef HAVE_FTRUNCATE
 int do_ftruncate(int fd, OFF_T size)
@@ -2423,18 +2243,6 @@ int do_mknod_atfd(int dfd, const char *name, mode_t mode, dev_t dev)
 #endif
 }
 
-int do_rename_atfd(int old_dfd, const char *old_name, int new_dfd, const char *new_name)
-{
-#ifdef AT_FDCWD
-	if (dry_run) return 0;
-	RETURN_ERROR_IF_RO_OR_LO;
-	return renameat(old_dfd, old_name, new_dfd, new_name);
-#else
-	(void)old_dfd; (void)old_name; (void)new_dfd; (void)new_name;
-	errno = ENOSYS;
-	return -1;
-#endif
-}
 
 #if defined HAVE_LINK || defined HAVE_LINKAT
 int do_link_atfd(int old_dfd, const char *old_name, int new_dfd, const char *new_name, int flags)
